+++ src/components/editorRemirror/appEditor.tsx
@@ -1,68 +1,21 @@
 import React from 'react';
-import { YjsExtension } from '@remirror/extension-yjs';
-import { AnnotationExtension } from '@remirror/extension-annotation';
 import { htmlToProsemirrorNode } from 'remirror';
-import { FindExtension } from '@remirror/extension-find';
-import Box from '@mui/material/Box';
-import data from 'svgmoji/emoji.json';
-import css from 'refractor/lang/css.js';
-import javascript from 'refractor/lang/javascript.js';
-import json from 'refractor/lang/json.js';
-import markdown from 'refractor/lang/markdown.js';
-import typescript from 'refractor/lang/typescript.js';
-import rust from 'refractor/lang/rust.js';
-import { cx } from '@remirror/core';
-import { TableComponents, TableExtension } from '@remirror/extension-react-tables';
+import { TableComponents } from '@remirror/extension-react-tables';
 import {
-  BlockquoteExtension,
-  BoldExtension,
-  BulletListExtension,
-  CalloutExtension,
-  CodeBlockExtension,
-  CodeExtension,
   ColumnAttributes,
-  ColumnsExtension,
-  createMarkPositioner,
-  DropCursorExtension,
-  EmojiExtension,
-  FontFamilyExtension,
-  FontSizeExtension,
-  HardBreakExtension,
-  HeadingExtension,
-  HistoryExtension,
-  ImageExtension,
-  ItalicExtension,
-  LinkExtension,
-  MentionAtomExtension,
-  NodeFormattingExtension,
-  OrderedListExtension,
-  ShortcutHandlerProps,
-  ShortcutsExtension,
-  StrikeExtension,
-  TaskListExtension,
-  TextColorExtension,
-  UnderlineExtension,
 } from 'remirror/extensions';
 import { AllStyledComponent } from '@remirror/styles/emotion';
-import { WebrtcProvider } from 'y-webrtc';
-import * as Y from 'yjs';
 import {
   CalloutTypeButtonGroup,
-  CommandButton,
   CommandButtonGroup,
-  CommandMenuItem,
   HeadingLevelButtonGroup,
   DecreaseFontSizeButton,
-  DropdownButton,
   EmojiPopupComponent,
-  FloatingToolbar,
-  FloatingWrapper,
   /* FindReplaceComponent, */
   HistoryButtonGroup,
   IncreaseFontSizeButton,
   IndentationButtonGroup,
   ListButtonGroup,
-  MentionAtomNodeAttributes,
   Remirror,
   TextAlignmentButtonGroup,
   ThemeProvider,
@@ -75,21 +28,15 @@ import {
   ToggleStrikeButton,
   ToggleUnderlineButton,
   Toolbar,
-  useActive,
-  useAttrs,
-  useChainedCommands,
-  useCommands,
-  useCurrentSelection,
-  useExtensionEvent,
-  useMentionAtom,
-  useUpdateReason,
   useRemirror,
   VerticalDivider,
 } from '@remirror/react';
-
-const room = 'test-placeholder-yjs';
-const ydoc = new Y.Doc();
-const provider = new WebrtcProvider(room, ydoc);
+import FontFamilyButtons from '@/editorRemirror/extensions/fontFamilyButtons';
+import FontSizeButtons from '@/editorRemirror/extensions/fontSizeButtons';
+import FloatingLinkToolbar from '@/editorRemirror/floatingLinkToolbar';
+import LineHeightButtonDropdown from './extensions/lineHeightButtonDropDown';
+import MentionSuggestor from '@/editorRemirror/extensions/mentionSuggestor';
+import extensions from './extensions';
 
 const TWO_COLUMNS: ColumnAttributes = {
   count: 2,
@@ -101,381 +48,6 @@ const THREE_COLUMNS: ColumnAttributes = {
   fill: 'balance',
   gap: '2em',
 };
-
-const FONT_FAMILIES: Array<[React.CSSProperties['fontFamily'], string]> = [
-  ['serif', 'Serif'],
-  ['sans-serif', 'San serif'],
-  ['cursive', 'Cursive'],
-  ['fantasy', 'Fantasy'],
-];
-
-const ALL_USERS = [
-  { id: 'joe', label: 'Joe' },
-  { id: 'sue', label: 'Sue' },
-  { id: 'pat', label: 'Pat' },
-  { id: 'tom', label: 'Tom' },
-  { id: 'jim', label: 'Jim' },
-];
-
-const ALL_TAGS = [
-  { id: 'cel', label: 'Celebrity' },
-  { id: 'ed', label: 'Education' },
-  { id: 'tech', label: 'Tech' },
-];
-
-const FONT_SIZES = ['8', '10', '12', '14', '16', '18', '24', '30'];
-
-const FontFamilyButtons = () => {
-  const { setFontFamily } = useCommands();
-  const active = useActive();
-  return (
-    <CommandButtonGroup>
-      <DropdownButton aria-label='Font family' icon='text'>
-        {FONT_FAMILIES.map(([fontFamily, label]) => (
-          <CommandMenuItem
-            key={fontFamily}
-            commandName='setFontFamily'
-            onSelect={() => setFontFamily(fontFamily as string)}
-            enabled={setFontFamily.enabled(fontFamily as string)}
-            active={active.fontFamily({ fontFamily })}
-            label={<span style={{ fontFamily }}>{label}</span>}
-          />
-        ))}
-      </DropdownButton>
-    </CommandButtonGroup>
-  );
-};
-
-const FontSizeButtons = () => {
-  const { setFontSize } = useCommands();
-  const { fontSize } = useActive();
-  return (
-    <DropdownButton aria-label='Set font size' icon='fontSize'>
-      {FONT_SIZES.map((size) => (
-        <CommandMenuItem
-          key={size}
-          commandName='setFontSize'
-          onSelect={() => setFontSize(size)}
-          enabled={setFontSize.enabled(size)}
-          active={fontSize({ size })}
-          label={size}
-          icon={null}
-          displayDescription={false}
-        />
-      ))}
-    </DropdownButton>
-  );
-};
-
-function useLinkShortcut() {
-  const [linkShortcut, setLinkShortcut] = React.useState<ShortcutHandlerProps | undefined>();
-  const [isEditing, setIsEditing] = React.useState(false);
-
-  useExtensionEvent(
-    LinkExtension,
-    'onShortcut',
-    React.useCallback(
-      (props) => {
-        if (!isEditing) {
-          setIsEditing(true);
-        }
-
-        return setLinkShortcut(props);
-      },
-      [isEditing],
-    ),
-  );
-
-  return { linkShortcut, isEditing, setIsEditing };
-}
-
-function useFloatingLinkState() {
-  const chain = useChainedCommands();
-  const { isEditing, linkShortcut, setIsEditing } = useLinkShortcut();
-  const { to, empty } = useCurrentSelection();
-
-  const url = (useAttrs().link()?.href as string) ?? '';
-  const [href, setHref] = React.useState<string>(url);
-
-  // A positioner which only shows for links.
-  const linkPositioner = React.useMemo(() => createMarkPositioner({ type: 'link' }), []);
-
-  const onRemove = React.useCallback(() => chain.removeLink().focus().run(), [chain]);
-  const updateReason = useUpdateReason();
-
-  React.useLayoutEffect(() => {
-    if (!isEditing) {
-      return;
-    }
-
-    if (updateReason.doc || updateReason.selection) {
-      setIsEditing(false);
-    }
-  }, [isEditing, setIsEditing, updateReason.doc, updateReason.selection]);
-
-  React.useEffect(() => {
-    setHref(url);
-  }, [url]);
-
-  const submitHref = React.useCallback(() => {
-    setIsEditing(false);
-    const range = linkShortcut ?? undefined;
-
-    if (href === '') {
-      chain.removeLink();
-    } else {
-      chain.updateLink({ href, auto: false }, range);
-    }
-
-    chain.focus(range?.to ?? to).run();
-  }, [setIsEditing, linkShortcut, chain, href, to]);
-
-  const cancelHref = React.useCallback(() => {
-    setIsEditing(false);
-  }, [setIsEditing]);
-
-  const clickEdit = React.useCallback(() => {
-    if (empty) {
-      chain.selectLink();
-    }
-
-    setIsEditing(true);
-  }, [chain, empty, setIsEditing]);
-
-  return React.useMemo(
-    () => ({
-      href,
-      setHref,
-      linkShortcut,
-      linkPositioner,
-      isEditing,
-      clickEdit,
-      onRemove,
-      submitHref,
-      cancelHref,
-    }),
-    [href, linkShortcut, linkPositioner, isEditing, clickEdit, onRemove, submitHref, cancelHref],
-  );
-}
-
-// eslint-disable-next-line react/prop-types
-const DelayAutoFocusInput = ({ autoFocus, ...rest }: React.HTMLProps<HTMLInputElement>) => {
-  const inputRef = React.useRef<HTMLInputElement>(null);
-
-  React.useEffect(() => {
-    if (!autoFocus) {
-      return;
-    }
-
-    const frame = window.requestAnimationFrame(() => {
-      inputRef.current?.focus();
-    });
-
-    // eslint-disable-next-line consistent-return
-    return () => {
-      window.cancelAnimationFrame(frame);
-    };
-  }, [autoFocus]);
-
-  return <input ref={inputRef} {...rest} />;
-};
-
-const FloatingLinkToolbar = () => {
-  const {
-    isEditing,
-    linkPositioner,
-    clickEdit,
-    onRemove,
-    submitHref,
-    href,
-    setHref,
-    cancelHref,
-  } = useFloatingLinkState();
-  const active = useActive();
-  const activeLink = active.link();
-  const { empty } = useCurrentSelection();
-
-  const handleClickEdit = React.useCallback(() => {
-    clickEdit();
-  }, [clickEdit]);
-
-  const linkEditButtons = activeLink ? (
-    <>
-      <ToggleBoldButton />
-      <ToggleItalicButton />
-      <ToggleStrikeButton />
-      <ToggleUnderlineButton />
-      <CommandButton
-        commandName='updateLink'
-        onSelect={handleClickEdit}
-        icon='pencilLine'
-        enabled
-      />
-      <CommandButton commandName='removeLink' onSelect={onRemove} icon='linkUnlink' enabled />
-    </>
-  ) : (
-    <>
-      <ToggleBoldButton />
-      <ToggleItalicButton />
-      <ToggleStrikeButton />
-      <ToggleUnderlineButton />
-      <CommandButton commandName='updateLink' onSelect={handleClickEdit} icon='link' enabled />
-    </>
-  );
-
-  return (
-    <>
-      {!isEditing && <FloatingToolbar>{linkEditButtons}</FloatingToolbar>}
-      {!isEditing && empty && (
-        <FloatingToolbar positioner={linkPositioner}>{linkEditButtons}</FloatingToolbar>
-      )}
-
-      <FloatingWrapper
-        positioner='always'
-        placement='bottom'
-        enabled={isEditing}
-        renderOutsideEditor
-      >
-        <DelayAutoFocusInput
-          style={{ zIndex: 20 }}
-          autoFocus
-          placeholder='Enter link...'
-          onChange={(event: React.ChangeEvent<HTMLInputElement>) => setHref(event.target.value)}
-          value={href}
-          onKeyPress={(event: React.KeyboardEvent<HTMLInputElement>) => {
-            const { code } = event;
-
-            if (code === 'Enter') {
-              submitHref();
-            }
-
-            if (code === 'Escape') {
-              cancelHref();
-            }
-          }}
-        />
-      </FloatingWrapper>
-    </>
-  );
-};
-
-const MentionSuggestor: React.FC = () => {
-  const [options, setOptions] = React.useState<MentionAtomNodeAttributes[]>([]);
-  const {
-    state,
-    getMenuProps,
-    getItemProps,
-    indexIsHovered,
-    indexIsSelected,
-  } = useMentionAtom({
-    items: options,
-  });
-
-  React.useEffect(() => {
-    if (!state) {
-      return;
-    }
-
-    const searchTerm = state.query.full.toLowerCase();
-    let filteredOptions: MentionAtomNodeAttributes[] = [];
-
-    if (state.name === 'tag') {
-      filteredOptions = ALL_TAGS.filter((tag) => tag.label.toLowerCase().includes(searchTerm));
-    } else if (state.name === 'at') {
-      filteredOptions = ALL_USERS.filter((user) => user.label.toLowerCase().includes(searchTerm));
-    }
-
-    filteredOptions = filteredOptions.sort().slice(0, 5);
-    setOptions(filteredOptions);
-  }, [state]);
-
-  const enabled = Boolean(state);
-
-  return (
-    <FloatingWrapper positioner='cursor' enabled={enabled} placement='bottom-start'>
-      <div {...getMenuProps()} className='suggestions'>
-        {enabled && options.map((user: { id: string, label: string }, index: number) => {
-          const isHighlighted = indexIsSelected(index);
-          const isHovered = indexIsHovered(index);
-
-          return (
-            <div
-              key={user.id}
-              className={cx('suggestion', isHighlighted && 'highlighted', isHovered && 'hovered')}
-              {...getItemProps({
-                item: user,
-                index,
-              })}
-            >
-              {user.label}
-            </div>
-          );
-        })}
-      </div>
-    </FloatingWrapper>
-  );
-};
-
-const LineHeightButtonDropdown = () => {
-  const { setLineHeight } = useCommands();
-  return (
-    <CommandButtonGroup>
-      <DropdownButton aria-label='Line height' icon='lineHeight'>
-        <CommandMenuItem
-          commandName='setLineHeight'
-          onSelect={() => setLineHeight(1)}
-          enabled={setLineHeight.enabled(1)}
-          label='Narrow'
-        />
-        <CommandMenuItem
-          commandName='setLineHeight'
-          onSelect={() => setLineHeight(2)}
-          enabled={setLineHeight.enabled(2)}
-          label='Wide'
-        />
-      </DropdownButton>
-    </CommandButtonGroup>
-  );
-};
-
-const extensions = () => [
-  new YjsExtension({ getProvider: () => provider, disableUndo: true }),
-  new AnnotationExtension(),
-  new BlockquoteExtension(),
-  new BoldExtension(),
-  new BulletListExtension(),
-  new CalloutExtension(),
-  new CodeBlockExtension({
-    supportedLanguages: [css, javascript, json, markdown, typescript, rust],
-  }),
-  new CodeExtension(),
-  new DropCursorExtension(),
-  new ColumnsExtension(),
-  new EmojiExtension({ data, moji: 'noto' }),
-  new FindExtension(),
-  new FontFamilyExtension(),
-  new FontSizeExtension({ defaultSize: '16', unit: 'px' }),
-  new HardBreakExtension(),
-  new HeadingExtension(),
-  new HistoryExtension(),
-  new ImageExtension({ enableResizing: true }),
-  new ItalicExtension(),
-  new LinkExtension({ autoLink: true }),
-  new OrderedListExtension(),
-  new TaskListExtension(),
-  new MentionAtomExtension({
-    matchers: [
-      { name: 'at', char: '@' },
-      { name: 'tag', char: '#' },
-    ],
-  }),
-  new NodeFormattingExtension(),
-  new ShortcutsExtension(),
-  new StrikeExtension(),
-  new TableExtension(),
-  new TextColorExtension(),
-  new UnderlineExtension(),
-];
 
 export default function appEditor() {
   const { manager, state, setState } = useRemirror({
