+++ src/components/editorRemirror/data/fontSizesData.tsx
@@ -1,0 +1,3 @@
+const FONT_SIZES = ['8', '10', '12', '14', '16', '18', '24', '30'];
+
+export default FONT_SIZES;
